# openapi.yaml
openapi: 3.0.3
info:
  title: RTO Compliance Hub API
  version: 1.0.0
  description: |
    REST API for RTO (Registered Training Organization) Compliance Management & Governance Platform.
    
    This platform provides comprehensive compliance tracking, governance documentation, training product management,
    professional development tracking, feedback collection, and resource management for RTOs.
    
    ## Features
    - **Governance**: Manage policies, procedures, and compliance documentation
    - **Training Products**: Track training courses, SOPs, and assessment materials
    - **Feedback Management**: Collect and analyze learner, employer, and industry feedback
    - **Professional Development**: Track staff credentials and PD requirements
    - **Resource Management**: Manage assets, infrastructure, and learning resources
    - **Complaints & Appeals**: Workflow-driven complaint handling
    - **Compliance Tracking**: Map standards to policies and evidence
    
    ## Authentication
    All authenticated endpoints require a valid JWT Bearer token in the Authorization header.
    Use the `/api/v1/auth/login` endpoint to obtain a token.
    
    ## Rate Limiting
    - Authentication endpoints: 5 requests per 15 minutes per IP
    - Password reset endpoints: 3 requests per hour per IP
    - General API endpoints: 100 requests per 15 minutes per IP
    - Webhook endpoints: 30 requests per minute
    
    ## Error Handling
...
    ```json
    {
...
      "status": 400,
...
    }
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/users:
    get:
      tags:
        - Users
      summary: List users
      description: Get paginated list of users with optional filters
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: department
          in: query
          schema:
            $ref: '#/components/schemas/Department'
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/Role'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/UserStatus'
        - name: q
          in: query
          description: Search query (name or email)
          schema:
            type: string
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
          
    post:
      tags:
        - Users
      summary: Create user
      description: Create a new user account
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /api/v1/policies:
    get:
      tags:
        - Policies
      summary: List policies
      description: Get paginated list of policies with optional filters
      operationId: listPolicies
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: standardId
          in: query
          description: Filter by standard ID
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/PolicyStatus'
      responses:
        '200':
          description: List of policies
          content:
            application/json:
              schema:
                type: object
                properties:
                  policies:
                    type: array
                    items:
                      $ref: '#/components/schemas/Policy'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                    
    post:
      tags:
        - Policies
      summary: Create policy
      description: Create a new policy document
      operationId: createPolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyCreate'
      responses:
        '201':
          description: Policy created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
...
  /api/v1/training-products:
    get:
      tags:
        - Training Products
      summary: List training products
      description: Get paginated list of training products
      operationId: listTrainingProducts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [Active, Inactive]
      responses:
        '200':
          description: List of training products
          content:
            application/json:
              schema:
                type: object
                properties:
                  trainingProducts:
                    type: array
                    items:
                      $ref: '#/components/schemas/TrainingProduct'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                    
    post:
      tags:
        - Training Products
      summary: Create training product
      description: Create a new training product
      operationId: createTrainingProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainingProductCreate'
      responses:
        '201':
          description: Training product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingProduct'
...
  /api/v1/pd:
    get:
      tags:
        - Professional Development
      summary: List PD items
      description: Get paginated list of professional development items
      operationId: listPDItems
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: userId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/PDStatus'
        - name: dueBefore
          in: query
          description: Filter items due before this date
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of PD items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PDItem'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                    
    post:
      tags:
        - Professional Development
      summary: Create PD item
      description: Create a new professional development item
      operationId: createPDItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PDItemCreate'
      responses:
        '201':
          description: PD item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PDItem'
...
  /api/v1/credentials:
    get:
      tags:
        - Credentials
      summary: List credentials
      description: Get list of credentials with optional filters
      operationId: listCredentials
      parameters:
        - name: userId
          in: query
          schema:
            type: string
        - name: expiresBefore
          in: query
          description: Filter credentials expiring before this date
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  credentials:
                    type: array
                    items:
                      $ref: '#/components/schemas/Credential'
  /api/v1/feedback:
    get:
      tags:
        - Feedback
      summary: List feedback
      description: Get feedback submissions with filters
      operationId: listFeedback
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/FeedbackType'
        - name: courseId
          in: query
          schema:
            type: string
        - name: trainerId
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
        - name: anonymous
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of feedback submissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  feedback:
                    type: array
                    items:
                      $ref: '#/components/schemas/Feedback'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
  /api/v1/assets:
    get:
      tags:
        - Assets
      summary: List assets
      description: Get paginated list of assets
      operationId: listAssets
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/AssetStatus'
        - name: location
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of assets
          content:
            application/json:
              schema:
                type: object
                properties:
                  assets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Asset'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                    
    post:
      tags:
        - Assets
      summary: Create asset
      description: Create a new asset record
      operationId: createAsset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetCreate'
      responses:
        '201':
          description: Asset created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
...
  /api/v1/complaints:
    get:
      tags:
        - Complaints
      summary: List complaints
      description: Get paginated list of complaints
      operationId: listComplaints
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ComplaintStatus'
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of complaints
          content:
            application/json:
              schema:
                type: object
                properties:
                  complaints:
                    type: array
                    items:
                      $ref: '#/components/schemas/Complaint'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                    
    post:
      tags:
        - Complaints
      summary: Create complaint
      description: Create a new complaint record
      operationId: createComplaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplaintCreate'
      responses:
        '201':
          description: Complaint created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Complaint'
...
  /api/v1/audit-logs:
    get:
      tags:
        - Audit Logs
      summary: List audit logs
      description: Get paginated list of audit log entries
      operationId: listAuditLogs
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: userId
          in: query
          schema:
            type: string
        - name: action
          in: query
          schema:
            type: string
        - name: entityType
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date-time
        - name: dateTo
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of audit logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLog'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token obtained from /api/v1/auth/login
  parameters:
...
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: perPage
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 30
  schemas:
...
    User:
      type: object
      required:
        - id
        - name
        - email
        - department
        - roles
        - status
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: John Smith
        email:
          type: string
          format: email
          example: john.smith@rto.com.au
        department:
          $ref: '#/components/schemas/Department'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        status:
          $ref: '#/components/schemas/UserStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserCreate:
      type: object
      required:
        - name
        - email
        - password
        - department
        - roles
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        department:
          $ref: '#/components/schemas/Department'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
...
    Department:
      type: string
      enum: [Training, Admin, Management, Support]
    Role:
      type: string
      enum: [Trainer, SystemAdmin, ComplianceAdmin, Manager, Staff]
    UserStatus:
      type: string
      enum: [Active, Inactive]
      
    # Policy schemas
    Policy:
      type: object
      required:
        - id
        - title
        - version
        - status
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: Student Welfare Policy
        version:
          type: string
          example: "2.1"
        ownerId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/PolicyStatus'
        reviewDate:
          type: string
          format: date
        fileUrl:
          type: string
          format: uri
        mappedStandardIds:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PolicyCreate:
      type: object
      required:
        - title
        - version
        - fileUrl
      properties:
        title:
          type: string
        version:
          type: string
        ownerId:
          type: string
          format: uuid
        reviewDate:
          type: string
          format: date
        fileUrl:
          type: string
          format: uri
...
    PolicyStatus:
      type: string
      enum: [Draft, Published, Archived]
...
    TrainingProduct:
      type: object
      required:
        - id
        - code
        - name
        - status
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          example: TLI41221
        name:
          type: string
          example: Certificate IV in Transport and Logistics
        status:
          type: string
          enum: [Active, Inactive]
        sopIds:
          type: array
          items:
            type: string
            format: uuid
        assessmentStrategyUrl:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
    TrainingProductCreate:
      type: object
      required:
        - code
        - name
      properties:
        code:
          type: string
        name:
          type: string
        assessmentStrategyUrl:
          type: string
          format: uri
          
    # SOP schemas
...
    PDItem:
      type: object
      required:
        - id
        - userId
        - title
        - status
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        title:
          type: string
          example: Attend TAE Training
        hours:
          type: number
          nullable: true
        dueAt:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/PDStatus'
        evidenceUrl:
          type: string
          format: uri
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    PDItemCreate:
      type: object
      required:
        - userId
        - title
        - dueAt
      properties:
        userId:
          type: string
          format: uuid
        title:
          type: string
        hours:
          type: number
        dueAt:
          type: string
          format: date-time
    PDStatus:
      type: string
      enum: [Planned, Due, Overdue, Completed, Verified]
      
    # Credential schemas
    Credential:
      type: object
      required:
        - id
        - userId
        - name
        - issuedAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
          example: Certificate IV in Training and Assessment
        issuedAt:
          type: string
          format: date
        expiresAt:
          type: string
          format: date
          nullable: true
        evidenceUrl:
          type: string
          format: uri
...
    Feedback:
      type: object
      required:
        - id
        - type
        - anonymous
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/FeedbackType'
        courseId:
          type: string
          nullable: true
        trainerId:
          type: string
          nullable: true
        rating:
          type: number
          minimum: 0
          maximum: 5
        comments:
          type: string
          nullable: true
        anonymous:
          type: boolean
        createdAt:
          type: string
          format: date-time
    FeedbackType:
      type: string
      enum: [learner, employer, industry]
      
    # Asset schemas
    Asset:
      type: object
      required:
        - id
        - type
        - name
        - status
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          example: Crane
        name:
          type: string
          example: Mobile Crane #1
        location:
          type: string
          example: Training Yard A
        status:
          $ref: '#/components/schemas/AssetStatus'
        lastServiceAt:
          type: string
          format: date
          nullable: true
        documents:
          type: array
          items:
            type: string
            format: uri
        createdAt:
          type: string
          format: date-time
    AssetCreate:
      type: object
      required:
        - type
        - name
      properties:
        type:
          type: string
        name:
          type: string
        location:
          type: string
...
    AssetStatus:
      type: string
      enum: [Available, Assigned, Servicing, Retired]
      
    # Complaint schemas
    Complaint:
      type: object
      required:
        - id
        - source
        - description
        - status
      properties:
        id:
          type: string
          format: uuid
        source:
          type: string
          example: Student
        description:
          type: string
        status:
          $ref: '#/components/schemas/ComplaintStatus'
        studentId:
          type: string
          nullable: true
        trainerId:
          type: string
          nullable: true
        courseId:
          type: string
          nullable: true
        rootCause:
          type: string
          nullable: true
        correctiveAction:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ComplaintCreate:
      type: object
      required:
        - source
        - description
      properties:
        source:
          type: string
        description:
          type: string
        studentId:
          type: string
        trainerId:
          type: string
        courseId:
          type: string
...
    ComplaintStatus:
      type: string
      enum: [New, InReview, Actioned, Closed]
      
    # Evidence schema
...
    AuditLog:
      type: object
      required:
        - id
        - userId
        - action
        - entityType
        - timestamp
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        action:
          type: string
          example: login_attempt
        entityType:
          type: string
          example: user
        entityId:
          type: string
          nullable: true
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        ipAddress:
          type: string
          nullable: true
          
    # Pagination schema
    Pagination:
      type: object
      required:
        - page
        - perPage
        - total
        - totalPages
      properties:
        page:
          type: integer
          example: 1
        perPage:
          type: integer
          example: 30
        total:
          type: integer
          example: 156
        totalPages:
          type: integer
          example: 6
          
    # Error schemas
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          example: https://tools.ietf.org/html/rfc7231#section-6.5.1
        title:
          type: string
          example: Bad Request
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: Validation failed for request body
        instance:
          type: string
          example: /api/v1/users
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  responses:
    BadRequestError:
      description: Bad Request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://tools.ietf.org/html/rfc7231#section-6.5.1
            title: Bad Request
            status: 400
            detail: Validation failed
    UnauthorizedError:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://tools.ietf.org/html/rfc7231#section-6.5.1
            title: Unauthorized
            status: 401
            detail: Invalid or expired authentication token
    ForbiddenError:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://tools.ietf.org/html/rfc7231#section-6.5.3
            title: Forbidden
            status: 403
            detail: Insufficient permissions to access this resource
...